+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file contains "important" commits, small ones will probably not be added.
 
+### 200620 - 1 [1.0.2]
+- New icons for opening and closing a terminal
+- Fixed some dropmenus were showinh "undefined" as it's hint
+
 ### 200619 - 2 [1.0.2]
 - Improved terminal looking
 - Fixed the welcome message wasn't centralized when a terminal was opened